Consolidated MovieDetails and ShowDetails into one folder and stylesheet.
