--- src/components/ShowDetails/ShowDetails.scss
@@ -1,119 +1,0 @@
-@import "../../styles/variables";
-
-.show-wrapper {
-  font-family: $primary-font;
-  color: $primary-font-color;
-  display: flex;
-  height: 750px;
-
-  .show-container {
-    margin: 62px 85px 62px 85px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 100%;
-  }
-}
-
-.show-poster {
-  display: flex;
-  height: 100%;
-
-  img {
-    max-width: 430px;
-    border-radius: 20px;
-    box-shadow: 6px 9px 19px 1px rgba(0, 0, 0, 0.54);
-  }
-}
-
-.show-text-container {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-evenly;
-  // background-color: blue;
-  margin: 0;
-  margin-left: 100px;
-
-  .show-title {
-    font-size: 2.6rem;
-    font-weight: 700;
-  }
-
-  .show-tagline {
-    font-size: 1.3rem;
-    font-style: italic;
-  }
-
-  .show-info {
-    margin-top: 0.5rem;
-    font-size: 1.05rem;
-  }
-
-  .show-rating-container {
-    display: flex;
-    align-items: center;
-    margin-top: 0.5rem;
-    margin-bottom: 2.1rem;
-
-    .show-rating-stars {
-      color: $secondary-color;
-      font-size: 1.5rem;
-    }
-
-    .show-rating-text {
-      font-size: 1.2rem;
-      padding-left: 7px;
-      font-weight: 700;
-    }
-  }
-
-  .show-secondary-title {
-    margin: 1.2rem 0;
-    font-size: 1.4rem;
-    font-weight: 700;
-  }
-
-  .show-synopsis {
-    font-size: 1rem;
-  }
-}
-
-.buttons-container {
-  margin-top: 1.5rem;
-  display: flex;
-  align-items: center;
-  height: 100%;
-  margin-bottom: 0.75rem;
-}
-
-.button1,
-.button2,
-.button3 {
-  background: none;
-  cursor: pointer;
-  outline: none;
-  border: 1px solid $secondary-color;
-  border-radius: 25px;
-  padding: 14px 25px 14px 25px;
-  margin-right: 2rem;
-  font-size: 1.13rem;
-  font-weight: 700;
-  text-decoration: none;
-  color: $primary-font-color;
-  i {
-    padding-right: 3px;
-  }
-  &:hover {
-    background: $secondary-color;
-    color: $secondary-font-color;
-    transition: 0.4s;
-    transform: translateY(-3px);
-  }
-  &:active {
-    transition: 0.2s;
-    transform: translateY(4px);
-  }
-  &:focus {
-    transition: 0.2s;
-  }
-}
